Refactor contribution serializer and validation
