+++ lib/contracts/contribution.js
@@ -1,20 +1,5 @@
 const ethers = require('ethers');
 
-const schemas = require('kosmos-schemas');
-const tv4 = require('tv4');
-const validator = tv4.freshApi();
-
-validator.addFormat({
-  'date': function(value) {
-    const dateRegexp = /^[0-9]{4,}-[0-9]{2}-[0-9]{2}$/;
-    return dateRegexp.test(value) ? null : "A valid ISO 8601 full-date string is expected";
-  },
-  'time': function(value) {
-    const timeRegexp = /^([01][0-9]|2[0-3]):([0-5][0-9]):([0-5][0-9]|60)(\.[0-9]+)?(([Zz])|([\+|\-]([01][0-9]|2[0-3]):[0-5][0-9]))$/;
-    return timeRegexp.test(value) ? null : "A valid ISO 8601 full-time string is expected";
-  }
-})
-
 const ContributionSerializer = require('../serializers/contribution');
 const Base = require('./base');
 
@@ -62,12 +47,13 @@ class Contribution extends Base {
       });
   }
 
-  addContribution(contributionAttr, callOptions = {}) {
-    let jsonStr = ContributionSerializer.serialize(contributionAttr);
+  async addContribution(contributionAttr, callOptions = {}) {
+    const contribution = new ContributionSerializer(contributionAttr);
 
-    // Validate JSON document against schema
-    let result = validator.validate(JSON.parse(jsonStr), schemas['contribution']);
-    if (!result) { return Promise.reject(validator.error); }
+    try { await contribution.validate(); }
+    catch (error) { return Promise.reject(error); }
+
+    const jsonStr = contribution.serialize();
 
     return this.ipfs
       .add(jsonStr)
